backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
requests==2.31.0
email-validator==2.1.0
python-dotenv==1.0.0
user-agents==2.2.0
httpx==0.25.0
google-auth==2.29.0
Pillow==10.0.0
